+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.3.0](https://github.com/mrhoribu/simu-rewards/compare/v1.2.4...v1.3.0) (2026-01-25)
+
+
+### Features
+
+* add release-please; ([a5d6b46](https://github.com/mrhoribu/simu-rewards/commit/a5d6b466b4e322cee120bef49fb3adf099f081a5))
+* add retry support via nick-fields/retry ([d30ac72](https://github.com/mrhoribu/simu-rewards/commit/d30ac724534057e42e99b6178340ca42b143d033))
+* add timeout logic; ([532aa0a](https://github.com/mrhoribu/simu-rewards/commit/532aa0a0ac5841b6a3bb101593dc2574ba955a18))
+
+
+### Bug Fixes
+
+* account.ts update getToken method to use cheerio API ([8a1172e](https://github.com/mrhoribu/simu-rewards/commit/8a1172e65b63b0b61f08904f1e79054c475b70cc))
+* account.ts update getToken method to use cheerio API ([#1](https://github.com/mrhoribu/simu-rewards/issues/1)) ([e7624ce](https://github.com/mrhoribu/simu-rewards/commit/e7624ceee1457a9735cc4b81c91468d1eec93c8f))
+* add checkout step; ([d406bee](https://github.com/mrhoribu/simu-rewards/commit/d406beedd4be1756b7f66c331b503d6d6862fbad))
+* add minor tags; ([682dbc7](https://github.com/mrhoribu/simu-rewards/commit/682dbc72b85e54b6f2b90cbda6bf3f25d7f629b8))
+* correctly delete major tags in cd; ([4b32067](https://github.com/mrhoribu/simu-rewards/commit/4b320678892a5cbecdf2fc1d5a73589f87f2b8ad))
+* gameCode reference; ([5c2d1d3](https://github.com/mrhoribu/simu-rewards/commit/5c2d1d341bf4693d5782ff4f0d4dacf14e9d9ee7))
+* master -&gt; main; ([5f2d800](https://github.com/mrhoribu/simu-rewards/commit/5f2d800e4a400d68334fd221420bc65d41de15ff))
+* use correct .git; ([0dfef9e](https://github.com/mrhoribu/simu-rewards/commit/0dfef9eae6c85876d557792bdd7fd841cdfa8db9))
+
 ## [1.2.4](https://github.com/elanthia-online/simu-rewards/compare/v1.2.3...v1.2.4) (2026-01-24)